Add reimbursement invoice creation flow and enhance reimbursement logic
- Introduce `ReimbursementInvoiceCreate` model with JSON derivations for encoding and decoding.
- Add `createReimbursementInvoice` method in `InvoiceServiceLive` for creating reimbursement invoices.
- Update `Invoice` and related models to include `isReimbursement` property.
- Refactor `drivers` field in `Invoice` to `driver` for single driver association, updating schema, migrations, and repository logic.
- Modify reimbursement logic and adopt `Long` for consistent type handling.
- Expand in-memory and domain tests with new cases for reimbursement invoices, ensuring accuracy in reimbursement calculations and storage validation.

+++ domain/src/scala/domain/models/invoice/Reimbursement.scala
@@ -6,8 +6,8 @@ import zio.json.*
 
 case class Reimbursement(
   driverName: DriverName,
-  totalAmount: Float,
-  to: Map[DriverName, Float]
+  totalAmount: Long,
+  to: Map[DriverName, Long]
 )
 
 object Reimbursement {
@@ -16,5 +16,5 @@ object Reimbursement {
 
   given Schema[Reimbursement] = Schema.derived[Reimbursement]
 
-  given Schema[Map[DriverName, Float]] = Schema.schemaForMap[DriverName, Float](DriverName.encode)
+  given Schema[Map[DriverName, Long]] = Schema.schemaForMap[DriverName, Long](DriverName.encode)
 }
